+++ src/app/leader/view/leader.view.component.ts
@@ -4,23 +4,28 @@ import { MdButton } from '@angular2-material/button/button';
 import { MdToolbar } from '@angular2-material/toolbar/toolbar';
 import { MdIcon, MdIconRegistry } from '@angular2-material/icon/icon';
 import { MD_GRID_LIST_DIRECTIVES } from '@angular2-material/grid-list/grid-list';
-import { ActivatedRoute } from '@angular/router';
+import { LeaderModel, LeaderService } from '../../shared/leader/index';
+import { ROUTER_DIRECTIVES, Router, ActivatedRoute } from '@angular/router';
 
 @Component({
   moduleId: module.id,
   templateUrl: './leader.view.component.html',
   styleUrls: ['./leader.view.component.css'],
-  directives: [MD_GRID_LIST_DIRECTIVES, MdCard, MdButton, MdIcon, MdToolbar],
-  providers: [MdIconRegistry]
+  directives: [ROUTER_DIRECTIVES, MD_GRID_LIST_DIRECTIVES, MdCard, MdButton, MdIcon, MdToolbar],
+  providers: [MdIconRegistry, LeaderService]
 })
 
 export class LeaderViewComponent {
+
+  leader: LeaderModel = new LeaderModel()
 
   /**
    * Dependency Injection: route (for reading params later)
    */
   constructor(
-    private route: ActivatedRoute
+    private router: Router,
+    private route: ActivatedRoute,
+    private leaderService: LeaderService
   ){}
 
   /**
@@ -31,8 +36,37 @@ export class LeaderViewComponent {
     this.route.params
       .map(params => params['id'])
       .subscribe((id) => {
-        console.log('Leader View, ID from route params is:', id)
-      });
+        console.log('View Leader by ID from route params:', id)
+        if (id) {
+          this.leaderService.getLeader(id)
+          .subscribe(
+            data => {
+              this.setLeader(data)
+            },
+            err => console.error(err),
+            () => {}
+          )
+        }
+      })
   }
 
+  /**
+   * Leader loading handler
+   * @param {data} Loaded leader data
+   */
+  setLeader(data){
+    this.leader = data;
+  }
+
+  /**
+   * Remove this leader
+   * @param {leader} Leader being viewed
+   */
+  private deleteLeader(leader: LeaderModel) {
+    // Delete from DB
+    this.leaderService.deleteLeader(leader)
+
+    this.router.navigate(['/leaders'])
+    return false;
+  }
 }