+++ src/app/shared/leader-list/index.ts
@@ -1,3 +1,0 @@
-export * from './leader-list.service';
-export * from './leader.service';
-export * from './leader.model';